pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "angie"
version = "0.1.0"
description = "Angie — 24/7 personal AI assistant"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    # Core AI
    "pydantic-ai>=0.0.14",
    "github-copilot-sdk>=0.1.0",
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "aiomysql>=0.2.0",
    "PyMySQL>=1.1.0",
    # Cache & Queue
    "redis>=5.0.0",
    "celery[redis]>=5.4.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Config
    "pydantic-settings>=2.3.0",
    # Templating
    "jinja2>=3.1.0",
    # CLI
    "click>=8.1.0",
    "rich>=13.7.0",
    # Scheduling
    "apscheduler>=3.10.0",
    # Channels
    "slack-sdk>=3.27.0",
    "discord.py>=2.3.0",
    "aiohttp>=3.9.0",
    # Agent SDKs
    "google-api-python-client>=2.130.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "O365>=2.0.35",
    "phue>=1.1",
    "spotipy>=2.23.0",
    "PyGithub>=2.3.0",
    "aiounifi>=76",
    # Utilities
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "tenacity>=8.3.0",
    "croniter>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pyinstaller>=6.6.0",
]

[project.scripts]
angie = "angie.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/angie"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
]
ignore = [
    "E501",   # line length (handled by formatter)
    "B008",   # do not perform function calls in default arguments
    "UP046",  # Generic[T] subclass pattern (SQLAlchemy compat)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
pythonpath = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
